+++ packages/jazz-svelte/src/lib/index.ts
@@ -1,4 +1,5 @@
 export { createInviteLink, parseInviteLink } from "jazz-browser";
 export * from "./auth/index.js";
 export * from "./jazz.svelte.js";
+export * from "./jazz.class.svelte.js";
 export { useIsAuthenticated } from "./auth/useIsAuthenticated.svelte.js";