+++ examples/file-share-svelte/CHANGELOG.md
@@ -1,5 +1,22 @@
 # file-share-svelte
 
+## 0.0.74
+
+### Patch Changes
+
+- Updated dependencies [ec546b4]
+  - jazz-svelte@0.13.24
+
+## 0.0.73
+
+### Patch Changes
+
+- Updated dependencies [3431076]
+- Updated dependencies [02a240c]
+  - jazz-svelte@0.13.23
+  - jazz-tools@0.13.23
+  - jazz-inspector-element@0.13.23
+
 ## 0.0.72
 
 ### Patch Changes