+++ examples/multi-cursors/CHANGELOG.md
@@ -1,5 +1,13 @@
 # multi-cursors
 
+## 0.0.83
+
+### Patch Changes
+
+- Updated dependencies [02a240c]
+  - jazz-tools@0.13.23
+  - jazz-react@0.13.23
+
 ## 0.0.82
 
 ### Patch Changes