+++ examples/multiauth/CHANGELOG.md
@@ -1,5 +1,14 @@
 # multiauth
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [02a240c]
+  - jazz-tools@0.13.23
+  - jazz-react@0.13.23
+  - jazz-react-auth-clerk@0.13.23
+
 ## 0.0.30
 
 ### Patch Changes