+++ examples/passkey/CHANGELOG.md
@@ -1,5 +1,13 @@
 # minimal-auth-passkey
 
+## 0.0.88
+
+### Patch Changes
+
+- Updated dependencies [02a240c]
+  - jazz-tools@0.13.23
+  - jazz-react@0.13.23
+
 ## 0.0.87
 
 ### Patch Changes