+++ examples/pets/CHANGELOG.md
@@ -1,5 +1,13 @@
 # jazz-example-pets
 
+## 0.0.207
+
+### Patch Changes
+
+- Updated dependencies [02a240c]
+  - jazz-tools@0.13.23
+  - jazz-react@0.13.23
+
 ## 0.0.206
 
 ### Patch Changes