+++ packages/jazz-inspector-element/CHANGELOG.md
@@ -1,5 +1,13 @@
 # jazz-inspector-element
 
+## 0.13.23
+
+### Patch Changes
+
+- Updated dependencies [02a240c]
+  - jazz-tools@0.13.23
+  - jazz-inspector@0.13.23
+
 ## 0.13.21
 
 ### Patch Changes